+++ nptyping/_hashed_subscriptable_type.py
@@ -1,0 +1,17 @@
+from typing import Any, Dict
+
+from typish import SubscriptableType
+
+
+class HashedSubscriptableType(SubscriptableType):
+    _cache: Dict[int, Any] = {}
+
+    def __getitem__(cls, item: Any) -> Any:
+        # Take the item's class name into account to distinguish between
+        # e.g. ints and floats.
+        hashed = hash('{}{}{}'.format(
+            cls.__name__, item.__class__.__name__, item))
+        if hashed not in cls._cache:
+            cls._cache[hashed] = SubscriptableType.__getitem__(cls, item)
+        result = cls._cache[hashed]
+        return result